move ml request to action
